# Conflicts:
#	app/src/main/java/com/mobileprism/fishing/di/KoinModules.kt
#	app/src/main/java/com/mobileprism/fishing/domain/NewCatchMasterViewModel.kt
#	app/src/main/java/com/mobileprism/fishing/model/datasource/WeatherRepositoryRetrofitImpl.kt
#	app/src/main/java/com/mobileprism/fishing/model/repository/app/WeatherRepository.kt
#	app/src/main/java/com/mobileprism/fishing/model/use_cases/GetNewCatchWeatherUseCase.kt
#	app/src/main/java/com/mobileprism/fishing/ui/home/new_catch/NewCatchWeather.kt
#	app/src/main/java/com/mobileprism/fishing/ui/home/new_catch/pages/NewCatchWeather.kt
#	app/src/main/java/com/mobileprism/fishing/viewmodels/MapViewModel.kt

+++ app/src/main/java/com/mobileprism/fishing/model/mappers/MapMarkerMapper.kt
@@ -12,8 +12,8 @@ class MapMarkerMapper {
         return UserMapMarker(
             id = getNewMarkerId(),
             userId = getCurrentUserId(),
-            latitude = newMarker.latitude.toDouble(),
-            longitude = newMarker.longitude.toDouble(),
+            latitude = newMarker.latitude,
+            longitude = newMarker.longitude,
             title = newMarker.title,
             description = newMarker.description,
             markerColor = newMarker.markerColor,
